Create ChickenParmesan Class
Please let me know if I am understanding this correctly but is this requesting a pull? By the way, I made a bunch of other foods, I'd just like some insight as to what else I'm missing on this ChickenParmesan one; such as getter and setter methods or does extending already do that. Also, field initialization.